Reorganize all redux-related things
Also fix all of the import statements related to said redux-related things
