+++ app/components/AdvocateLogin.jsx
@@ -4,7 +4,7 @@ import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { browserHistory } from 'react-router'
 
-import { login } from '../reducers/auth'
+import { login } from '../redux/reducers/auth'
 
 /* ----------------- COMPONENT ------------------ */
