+++ app/main.jsx
@@ -3,11 +3,11 @@ import React from 'react'
 import { Router, Route, IndexRedirect, browserHistory } from 'react-router'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
-import { receiveIssues, receiveAdvocateIssues, receiveOpenIssues } from './reducers/issues'
-import { receiveVolunteers } from './reducers/volunteers'
-import { receiveAdvocates } from './reducers/advocates'
+import { receiveIssues, receiveAdvocateIssues, receiveOpenIssues } from './redux/reducers/issues'
+import { receiveVolunteers } from './redux/reducers/volunteers'
+import { receiveAdvocates } from './redux/reducers/advocates'
 
-import store from './store'
+import store from './redux/store'
 import App from './components/App'
 import Landing from './components/Landing'
 import Hotline from './components/Hotline'
