+++ app/redux/reducers/issues.jsx
